refactor: isolate on win action
