+++ random-game/styles/components/_popup.scss
@@ -22,7 +22,7 @@
   &_hidden {
     opacity: 0;
     transition: opacity 0.3s;
-    z-index: 0;
+    z-index: -1;
   }
 }
